all: Support of two byte UTF-8 characters on keyboard input
Also: change the handling of keyboard input througout the code.
Key functions and character input are now seperately handled, and
all occurences of bytes data type for text are removed.
This results in pretty many changed lines with marginal changes each.
Also: Guess on reading a file whther it should be rewritten with tabs