pye.py, pye2.py: Change home/end key behaviour
At the first key press of home or end key, the cursor is
moved to the start resp. end of the line. At the second key press, ist
goes to start-of-text resp end-of-text before trailing comments.

Also: When opening pye w/o a filename, it will show a window with the
file sin the current directory.