pye: Change the method for detecting file changes
The previous method for telling, whether a file was changed, was
unreliable. This commit changes it to a slightly better version by
comparing hash value over the content. Since the hash calculation is
rather slow, it is not done on every keypress, but only on file load,
file save, quit, redraw by Ctrl-E, undo and redo.