pye_xx: Size reduced version of pye.py
By removing some of it's features:
- Mouse support
- Bracket match
- Word skip & word delete
- Block comments
- Shifting lines up/down with Alt-up/down
- Writing tabs on saving
- Drop Regular expressions on search
- Simpler line edit (no word paste from text)

The character-wise mark/cut/copy/paste is maintained, as well
as Undo and Redo. While the latter may not be so important, it uses
the same code as Undo. So dropping it is complicated.
The size of the pre-compiled file drops by ~4k to ~12k