First time for version 2.72 file for online help
Organized geographically, that is, commands that reference the top of the file are at the top of the help.

Use ^o h  to open this help file In another window, switch windows with ^w. When help is needed just ^w again.

No source code changes are necessary and the memory it requires is optional, that is, if you don't need it don't open it.
If you want different or additional notes just edit the file h.